Refactor typing (#80)

+++ app/schemas/requests/bodies/product_review.py
@@ -10,4 +10,4 @@ from ...schema import ApplicationSchema
 class ProductReview(ApplicationSchema):
     product_review_photo: Optional[List[HttpUrl]] = None
     product_review_text: str
-    product_review_grade: int
+    product_review_grade: float
